docs(stories): paypalScriptProvider split source code
