+++ src/stories/payPalScriptProvider/PayPalScriptProvider.stories.tsx
@@ -1,20 +1,19 @@
 import React, { FC, ReactElement } from "react";
 import type { PayPalScriptOptions } from "@paypal/paypal-js/types/script-options";
 import { action } from "@storybook/addon-actions";
-import type { Story } from "@storybook/react";
 
-import { getOptionsFromQueryString } from "./utils";
+import { getOptionsFromQueryString } from "../utils";
 import {
     PayPalScriptProvider,
     DISPATCH_ACTION,
     SCRIPT_LOADING_STATE,
-} from "../index";
+} from "../../index";
 import {
     getScriptID,
     destroySDKScript,
-} from "../context/scriptProviderContext";
-import { usePayPalScriptReducer } from "../hooks/scriptProviderHooks";
-import { generateDocPageStructure } from "./commons";
+} from "../../context/scriptProviderContext";
+import { usePayPalScriptReducer } from "../../hooks/scriptProviderHooks";
+import overrideStories from "./code";
 
 const scriptProviderOptions: PayPalScriptOptions = {
     "client-id": "test",
@@ -120,93 +119,4 @@ export const Default: FC<{ deferLoading: boolean }> = ({ deferLoading }) => {
     );
 };
 
-const getDefaultCode = (): string =>
-    `import {
-	PayPalScriptProvider,
-	usePayPalScriptReducer,
-	getScriptID,
-	destroySDKScript,
-} from "@paypal/react-paypal-js";
-
-const SCRIPT_PROVIDER_OPTIONS = {
-	"client-id": "test",
-};
-
-// Custom loader component
-const LoadScriptButton = () => {
-	const [{ isResolved }, dispatch] = usePayPalScriptReducer();
-
-	return (
-		<div style={{ display: "inline-flex" }}>
-			<button
-				type="button"
-				style={{ display: "block", marginBottom: "20px" }}
-				disabled={isResolved}
-				onClick={() => {
-					dispatch({
-						type: "setLoadingStatus",
-						value: "pending",
-					});
-				}}
-			>
-				Load PayPal script
-			</button>
-			<button
-				type="button"
-				style={{
-					display: "block",
-					marginBottom: "20px",
-					marginLeft: "1em",
-				}}
-				onClick={() => {
-					destroySDKScript(getScriptID(SCRIPT_PROVIDER_OPTIONS));
-					dispatch({
-						type: "setLoadingStatus",
-						value: "initial",
-					});
-				}}
-			>
-				Reset
-			</button>
-		</div>
-	);
-};
-
-// Show state
-function PrintLoadingState() {
-    const [{ isInitial, isPending, isResolved, isRejected }] =
-        usePayPalScriptReducer();
-    let status = "no status";
-
-	if (isInitial) {
-		status = "initial";
-	} else if (isPending) {
-		status = "pending";
-	} else if (isResolved) {
-		status = "resolved";
-	} else if (isRejected) {
-		status = "rejected";
-	}
-
-	return (<div>Current status: { status }</div>);
-}
-
-export default function App() {
-	return (
-		<PayPalScriptProvider
-			options={SCRIPT_PROVIDER_OPTIONS}
-			deferLoading={true}
-		>
-			<LoadScriptButton />
-			<PrintLoadingState />
-			{/* add your paypal components here (ex: <PayPalButtons />) */}
-		</PayPalScriptProvider>
-	);
-}
-`;
-
-(Default as Story).parameters = {
-    docs: {
-        page: () => generateDocPageStructure(getDefaultCode()),
-    },
-};
+overrideStories();
